spacemacs: Add initial support
This patch support spacemacs configuration files. Currently only private
layers and .spacemacs files are save. Probably that was is required!

The powerline indicator and some semantic settings are not applied
properly in the first line. This need to investigated and fixed later!
Work around is to run M-m f e R after the launch.

Currently following custom layers are supported
* my-common-setup
* my-helm-setup
* my-progmode-setup
* my-c-mode-setup
* my-gitmode-setup
* my-orgmode-setupmy-common-setup

Signed-off-by: Ratheesh S <ratheeshreddy@gmail.com>