vim: enable bracket paste mode
Bracketed paste was introduced in ViM-8.0.210 onwards. This will allow
underlying application to know if the pasted text came from user(shift-enter0)
or from inside application.

In ViM context, this commit will set paste mode automatically on
shift-enter so that it eliminate necessity to manually enable and
disable paste mode to avoid indentation issues. This works in normal,
visual and insert modes.

This should be created as a separate plugin in the future.

This code was copied from:
https://stackoverflow.com/questions/5585129/pasting-code-into-terminal-window-into-vim-on-mac-os-x

Signed-off-by: Ratheesh S <ratheeshreddy@gmail.com>
Tested-by: Ratheesh S <ratheeshreddy@gmail.com>