vim:wilder: Add support for completing history items only
This commit supports completing from history items from respective mode.
It is invokes using usual <C-f>. For more readability a star mark is
added to airline a section when plugin is in history completion mode.

PS: we need to disable <BS> key in normal mode as it is need for this
configuration. May need to reconsider to map this key to something else
in the future.

Discussion URL: https://github.com/gelguy/wilder.nvim/issues/14

Signed-off-by: Ratheesh S <ratheeshreddy@gmail.com>